Added Custom Keybinds again this time on my own branch
i have time to rant today so what shall we talk about? ooo F1 that was good today, boring ass redbull 1,2 but the madlad lando p3 which was pog. le clair got fked on the same corner he came off in quali which was dumb af. ooo also the LCS finals was today a quick 3-0 for EG which is eh but they did do the same thing as G2 which is intresting since they are in the same bracket as eachother for MSI so thats spicy but G2's real enemy is T1 with Faker the beast. erm what else.... pfft i dunno got any good series to watch? still waiting on moon knight new ep and halo but its weekly which is lame. yeah thats about it from me this time. cya next time.